refactor(gateway): split agent runtime helpers
